Move lightboxref shortcode into its own file
